+++ driver-core/src/main/java/com/datastax/driver/core/Cluster.java
@@ -1168,6 +1168,8 @@ public class Cluster implements Closeable {
 
             this.metrics = configuration.getMetricsOptions() == null ? null : new Metrics(this);
             this.listeners = new CopyOnWriteArraySet<Host.StateListener>(listeners);
+
+            this.scheduledTasksExecutor.scheduleWithFixedDelay(new CleanupIdleConnectionsTask(), 10, 10, TimeUnit.SECONDS);
         }
 
         // Initialization is not too performance intensive and in practice there shouldn't be contention
@@ -2224,6 +2226,19 @@ public class Cluster implements Closeable {
                         }
                     }
                 }).start();
+            }
+        }
+
+        private class CleanupIdleConnectionsTask implements Runnable {
+            @Override public void run() {
+                try {
+                    long now = System.currentTimeMillis();
+                    for (SessionManager session : sessions) {
+                        session.cleanupIdleConnections(now);
+                    }
+                } catch (Exception e) {
+                    logger.warn("Error while trashing idle connections", e);
+                }
             }
         }
     }